# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smirk-backend-core",
    "description": "Open, self-hostable backend for the Smirk non-custodial multi-chain wallet.",
    "contact": {
      "name": "John Winter Murphy",
      "email": "jw@such.software"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/keys": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register or update one of the authenticated user's per-asset public keys.",
        "description": "Idempotent: the DB upsert keys on `(user_id, asset, key_type)`, so re-posting\nthe same asset replaces the stored key material in place.",
        "operationId": "register_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Key registered or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserKeyInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid asset"
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/users/by-username/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Resolve a username to its user and that user's receiving public keys.",
        "description": "Public: this is how a sender discovers where to send. The lookup is\ncase-insensitive (the handle is canonicalized lowercase). An unknown handle\nreturns the same constant-shape response with `registered: false`.",
        "operationId": "lookup_username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username to resolve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Username resolution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupUsernameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/username": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "The authenticated user's own username (or `null`).",
        "operationId": "get_my_username",
        "responses": {
          "200": {
            "description": "Caller's username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyUsernameResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Claim or update the authenticated user's username.",
        "description": "The submitted value is lowercased, validated, and checked against the\nreserved list before the DB write. The UNIQUE constraint is the atomic claim:\na collision (taken or reserved) is a 409, never a 500.",
        "operationId": "set_username",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUsernameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Username claimed or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetUsernameResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "409": {
            "description": "Username is taken or reserved"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/{user_id}/keys": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List a user's per-asset receiving public keys.",
        "description": "Public: a sender fetches these to construct a payment to `user_id`.",
        "operationId": "get_user_keys",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user's public keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserKeysResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed user id"
          }
        }
      }
    },
    "/users/{user_id}/keys/{asset}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Fetch a user's receiving public key for one asset.",
        "operationId": "get_user_key_for_asset",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset",
            "in": "path",
            "description": "Asset (btc, ltc, xmr, wow, grin)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user's key for this asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserKeyInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid asset"
          },
          "404": {
            "description": "User has no key for this asset"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LookupUsernameResponse": {
        "type": "object",
        "description": "Result of resolving a username. Constant-shape: the same fields are always\npresent, so the response is not a structure oracle.",
        "required": [
          "registered"
        ],
        "properties": {
          "public_keys": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublicKeysInfo",
                "description": "The user's per-asset receiving public keys, if registered."
              }
            ]
          },
          "registered": {
            "type": "boolean",
            "description": "Whether a user owns this username."
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The resolved user id (UUID string), if registered."
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "description": "The canonical (lowercased) username, if registered."
          }
        }
      },
...
      "MyUsernameResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": [
              "string",
              "null"
            ],
            "description": "The caller's username, or `null` if they have not claimed one."
          }
        }
      },
...
      "PublicKeysInfo": {
        "type": "object",
        "description": "A user's per-asset receiving public keys, by asset.",
        "properties": {
          "btc": {
            "type": [
              "string",
              "null"
            ]
          },
          "grin": {
            "type": [
              "string",
              "null"
            ]
          },
          "ltc": {
            "type": [
              "string",
              "null"
            ]
          },
          "wow": {
            "type": [
              "string",
              "null"
            ]
          },
          "xmr": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RegisterKeyRequest": {
        "type": "object",
        "description": "Register or update one of the caller's per-asset public keys.",
        "required": [
          "asset",
          "public_key"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "Asset the key is for (`btc`, `ltc`, `xmr`, `wow`, `grin`)."
          },
          "public_key": {
            "type": "string",
            "description": "The public key (format depends on the asset)."
          },
          "public_spend_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "XMR/WOW only: the public spend key."
          }
        }
      },
...
      "SetUsernameRequest": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "Desired username (3-32 chars, lowercase `[a-z0-9_]`, no leading/trailing\nunderscore). Submitted values are lowercased before validation."
          }
        }
      },
      "SetUsernameResponse": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "The username as stored (lowercased)."
          }
        }
      },
...
      "UserKeyInfo": {
        "type": "object",
        "description": "A single per-asset public key.",
        "required": [
          "asset",
          "public_key"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "Asset the key is for (lowercase: `btc`, `ltc`, `xmr`, `wow`, `grin`)."
          },
          "public_key": {
            "type": "string"
          },
          "public_spend_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "XMR/WOW only: the public spend key."
          }
        }
      },
      "UserKeysResponse": {
        "type": "object",
        "description": "A user's full set of per-asset public keys.",
        "required": [
          "keys"
        ],
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserKeyInfo"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session access token issued by the /auth/* endpoints."
      }
    }
  },
...
}